feat(utxo-lib): make psbt taproot pub keys parsing a function
Later, it can be reused for tx to psbt conversion.

TICKET: BG-58455
