+++ modules/utxo-lib/src/bitgo/UtxoPsbt.ts
@@ -1,25 +1,14 @@
 import { Psbt as PsbtBase } from 'bip174';
 import { TapBip32Derivation, Transaction as ITransaction, TransactionFromBuffer } from 'bip174/src/lib/interfaces';
 import { checkForInput } from 'bip174/src/lib/utils';
-import * as opcodes from 'bitcoin-ops';
 import { BufferWriter, varuint } from 'bitcoinjs-lib/src/bufferutils';
 
-import {
-  taproot,
-  HDSigner,
-  Psbt,
-  PsbtTransaction,
-  Transaction,
-  TxOutput,
-  Network,
-  ecc as eccLib,
-  script as bscript,
-} from '..';
+import { taproot, HDSigner, Psbt, PsbtTransaction, Transaction, TxOutput, Network, ecc as eccLib } from '..';
 import { UtxoTransaction } from './UtxoTransaction';
 import { getOutputIdForInput } from './Unspent';
 import { isSegwit } from './psbt/scriptTypes';
 import { unsign } from './psbt/fromHalfSigned';
-import { toXOnlyPublicKey } from './outputScripts';
+import { parseTaprootScript2of3PubKeys, toXOnlyPublicKey } from './outputScripts';
 
 export interface HDTaprootSigner extends HDSigner {
   /**
@@ -170,17 +159,7 @@ export class UtxoPsbt<Tx extends UtxoTransaction<bigint>> extends Psbt {
     }
     const { controlBlock, script } = input.tapLeafScript[0];
     const witness: Buffer[] = [script, controlBlock];
-    const decompiled = bscript.decompile(script);
-    if (!decompiled || decompiled?.length !== 4) {
-      throw new Error('Not a valid bitgo n-of-n script.');
-    }
-    const [pubkey1, op_checksigverify, pubkey2, op_checksig] = decompiled;
-    if (!Buffer.isBuffer(pubkey1) || !Buffer.isBuffer(pubkey2)) {
-      throw new Error('Public Keys are not buffers.');
-    }
-    if (op_checksigverify !== opcodes.OP_CHECKSIGVERIFY || op_checksig !== opcodes.OP_CHECKSIG) {
-      throw new Error('Opcodes do not correspond to a valid bitgo script');
-    }
+    const [pubkey1, pubkey2] = parseTaprootScript2of3PubKeys(script);
     for (const pk of [pubkey1, pubkey2]) {
       const sig = input.tapScriptSig?.find(({ pubkey }) => pubkey.equals(pk));
       if (!sig) {
